chore: Update GitHub Actions workflow to set up GitLab user configuration
This commit adds a new step to the GitHub Actions workflow to set up the GitLab user configuration. The user name is set to "Lucas04" and the email is set to "lucas04@foxmail.com". This step ensures that the correct user information is used when interacting with GitLab during the synchronization process.